requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-multipart==0.0.9
chromadb==0.5.5
sentence-transformers==3.0.1
pdfminer.six==20231228
PyMuPDF==1.24.9
pytesseract==0.3.13
scikit-learn==1.5.1
